+++ js/constant.js
@@ -30,6 +30,8 @@ const START_POINTS_OBJECT = {
   lng: TOKYO_LONGITUDE,
 };
 
+const ZOOM = 13;
+
 const MODAL_SHOW_TIME = 5000;
 
 const ERROR_MESSAGE = 'Не удалось отправить форму. Попробуйте ещё раз';
@@ -38,5 +40,5 @@ export {
   MIN_ELEMENT, COUNT_OF_PINS, MIN_LOCATION_X, MIN_LOCATION_Y,
   MAX_LOCATION_X, MAX_LOCATION_Y, MAX_PRICE, MAX_COUNT_OF_AVATARS, MAX_COUNT_OF_DECIMAL_NUMBERS,
   MAX_FEATURES, MAX_PHOTOS, MIN_PRICE, TOKYO_LATITUDE, TOKYO_LONGITUDE, MAIN_PIN, PIN, START_POINTS,
-  START_POINTS_OBJECT, ERROR_MESSAGE, MODAL_SHOW_TIME
+  START_POINTS_OBJECT, ERROR_MESSAGE, MODAL_SHOW_TIME, ZOOM
 };
